fix problem with Windows build
fixes issue with json config file not parsing correctly. (I was using vmargs '-Xmx512m' and for some reason after parsing it was coming out blank.)

This patch updates the premake script for windows. The script now creates a static runtime, so there should be no need to install the C++ distributable.